+++ src/daemon/__tests__/http-server-tenant-trust.test.ts
@@ -5,7 +5,7 @@ import path from 'node:path';
 import { createDaemonHttpServer } from '../server/http-server.ts';
 import { resolveSessionRequestLogPath } from '../session-store.ts';
 import { safeSessionName } from '../session-paths.ts';
-import { DAEMON_HTTP_TENANT_HEADER } from '../http-contract.ts';
+import { DAEMON_HTTP_TENANT_HEADER } from '@agent-device/contracts/daemon-http';
 import type { DaemonRequest, DaemonResponse } from '../types.ts';
 import {
   closeLoopbackServer,
