--- src/daemon/http-contract.ts
@@ -1,28 +1,0 @@
-export const DAEMON_HTTP_BASE_PATH = '/agent-device';
-export const DAEMON_HTTP_TENANT_HEADER = 'x-agent-device-tenant';
-export const DAEMON_HTTP_NETWORK_ACCESS_HEADER = 'x-agent-device-network-access';
-export const DAEMON_HTTP_PUBLIC_NETWORK_ACCESS = 'public-only';
-
-export function buildDaemonHttpBaseUrl(baseUrl: string): string {
-  return buildDaemonHttpUrl(baseUrl, DAEMON_HTTP_BASE_PATH);
-}
-
-export function buildDaemonHttpUrl(baseUrl: string, route: string): string {
-  const normalizedBase = baseUrl.endsWith('/') ? baseUrl : `${baseUrl}/`;
-  return new URL(route.replace(/^\/+/, ''), normalizedBase).toString();
-}
-
-export function buildDaemonHttpAuthHeaders(token: string | undefined): Record<string, string> {
-  const normalizedToken = token?.trim();
-  if (!normalizedToken) return {};
-  return {
-    authorization: `Bearer ${normalizedToken}`,
-    'x-agent-device-token': normalizedToken,
-  };
-}
-
-export function buildDaemonHttpTenantHeaders(tenantId: string | undefined): Record<string, string> {
-  const normalizedTenantId = tenantId?.trim();
-  if (!normalizedTenantId) return {};
-  return { [DAEMON_HTTP_TENANT_HEADER]: normalizedTenantId };
-}
